Fix display errors
The display seems to be creating errors when the SPI interface can't correctly read commands. The solution is to create a service script to detect if the display encounters this issue, and reboots to reset the display since it becomes unresponsive. The frame rate and spi frequency have been updated to potentially reduce this data read/write error since the transfer rate may just be too high.