Reorganize pre- and post-reboot scripts
Pre- and post-reboot scripts were re-organized depending on whether
a script's task required a reboot in order for its changes to take
effect. The only operations that require a reboot are:

- Installing the WiFi drivers
- Setting the device tree overlays

The rest of operations and tasks do not require such additional
measures in order for their changes to take effect. Technically,
when creating system services, it is often safer to enable them
rather than activate them immediately after creation. This is
mainly because they start on a clean state.

Additionally, the scripts were numbered depending on the order that
they should be executed. For the most part, scripts can run in any
order as most of their operations are independent from each other.
However, they are some scripts that are dependent on the successful
execution of other scripts (for instance, installing the WiFi
drivers before connecting to WiFi). That said, the order in which
the scripts were ordered was based around this principle.