Commit 3_set_device_id script
This script will be executed pre-reboot, as a system reboot is
required for changes to take effect. In a nutshell, the script
first fetches teh last device ID in the "devices" table, parses it,
and generates a new device ID. Lastly, the newly-generated ID is
inserted into the table. The format of the device ID is "fpts-xxx".

Note: This script will most likely not work when scaling up the
system due to race conditions. Multiple devices might have the same
device ID but fail when trying to insert it into the "devices"
table.