Refactor get_new_device_id to fetch first unallocated device ID
Logic in get_new_device_id() was changed to fetch the first available
(unallocated) device ID from the database instead of generating a new
ID based on the last record's ID. This approach allows for repurposing
device IDs that were once taken but have been freed by a device.

In order to enable this behavior, an 'allocated' column was created in
the 'devices' table. This column serves to identify whether a particular
device ID was allocated. With this in mind, insert_device_id() was
modified to include an 'allocated' field in the database insert query,
marking the device as allocated upon insertion.

As of this commit, this implementation does not handle race conditions.