Modify Reading and Writing of Touch Sensor to Screen Pipe
Data being written to the pipe by the touch sensor is flushed
immediately to the pipe. This presumably might solve the issue of
missing taps. In theory, if the screen module reads from the pipe before
taps are flushed, then these taps are missed. Additionally, a newline is
also written into the pipe for readability purposes.

It is worth mentioning that threading must be implemented to read and
write to the pipe. These subprocesses will run indefinitely until the
system shutsdown. The processing of the taps will be done by the parent
process. A queue is an approprite approach for communicating between
parent and child processes.