Bugfix: Network Error Causing touch_sensor.py to Crash (#28)
* Logging Fix

* Update logger filenames

Need to remove the filename extension '.py' since the period is interpretted as a parent-child logger relationship.

* Fix logging file config

The config should not disable existing loggers when it is called since each script runs the file config.

* Add comment for suppressing loggers

* Modify Reading and Writing of Touch Sensor to Screen Pipe

Data being written to the pipe by the touch sensor is flushed
immediately to the pipe. This presumably might solve the issue of
missing taps. In theory, if the screen module reads from the pipe before
taps are flushed, then these taps are missed. Additionally, a newline is
also written into the pipe for readability purposes.

It is worth mentioning that threading must be implemented to read and
write to the pipe. These subprocesses will run indefinitely until the
system shutsdown. The processing of the taps will be done by the parent
process. A queue is an approprite approach for communicating between
parent and child processes.

* Fix logger reference in screen_utils.py

The logs in screen_util.py were done via 'logging' which is incorrect since that uses the root logger, and so the defined 'logger' should be used instead.

* Fix third-party logger suppression

* Fix log typo

* Add threading to screen

Because the single thread approach blocks during the drawing of popups, the screen misses any incoming taps during this period. The dual thread approach aims to fix this.

* Catch Network Error

When there is a network error, the touch sensor script crashes due to a thrown error, so we need to skip the request for now until we implement a local db.

---------

Co-authored-by: Juan Becerra <jdb9056@nyu.edu>
Co-authored-by: kc34251 <kc4433@nyu.edu>