improvement: Commit Global Variables for Shell Scripts
The motivation for this was to allow other shell scripts to access
shared resources, such as the color variables. This global file is
incomplete, as there are other shared globals that have not been added
to this file.

A separate ticket must be created to properly go through the shell
scripts, identify globals and properly move them to `globals.sh`. Any
script that requires these resources must source the script.