feature: Support Checking for Network Connection
Updating the device requires an active network connection in order to
pull the latest changes from the remote repository. This commit
introduces a network connection check before initiating the update
procedure. If the user is not connected to network, the script offers
the user the ability to attempt connecting. If the connection was not
successfully established after three attempts, the script exits.