feature: Support Updating Service Files with Templates
By design, this function does not return any error values as the status
of the services will be checked once the changes have been applied. In
case something goes wrong during the update procedure, the service files
will be invalid and, therefore, the services won't be able to start.

The only potential issue would be if `cp` fails to copy the new service
file to the original. However, any issues will be reported to the
screen, allowing us to identify whether an error occurred during the
process.

Lastly, after all service files are properly updates, the changes are
applied by reloading the daemon and each service is restarted.