feature: Support Checking Status of System Services
This is the last step of the update process. If at least one service is
down, the whole system is rolled back to the previous commit, improving
the system's resiliency.

Minor issues were addressed, syntax errors. Additionally, the CRITICAL
error message was added to `globals.sh` to provide for an additional
level of error severity.