Add reset count in write_device_state
If the device has yet to update its time due to lack of network connection and runs the application, it will not reset the unit count since the current time hasn't changed. The solution is to equip write_device_state with the same logic so that it continuously checks to see if it needs to reset the unit count.